feat: Added icons for different flavours
